requirements.txt
google-generativeai==0.3.1
python-dotenv
fastapi
pydantic
uvicorn